refactor: intergrate menu and context into a single file
